backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
asyncmy==0.2.10
aiomysql==0.2.0
mysql-connector-python==9.1.0

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
email-validator==2.2.0

# NLP (Faz 2 icin opsiyonel; semantic arama ilk kullanima kadar lazy import edilir)
# sentence-transformers==3.3.1
# torch==2.5.1

# Vector DB (Faz 2)
qdrant-client==1.12.1

# HTTP
httpx==0.28.1

# Import scripti için
tqdm==4.67.1
